setup crud for child category completely and set delete protection for category relations

+++ app/Http/Controllers/Backend/SubCategoryController.php
@@ -6,6 +6,7 @@ use App\Models\Category;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\DataTables\SubCategoryDataTable;
+use App\Models\ChildCategory;
 use App\Models\SubCategory;
 use Str;
 
@@ -96,6 +97,10 @@ class SubCategoryController extends Controller
     public function destroy(string $id)
     {
         $subCategory = SubCategory::findOrFail($id);
+        $childCategory = ChildCategory::where('sub_category_id' , $subCategory->id)->count();
+        if($childCategory > 0){
+            return response(['status' => 'error' , 'message' => 'This Sub Category contains, sub items for delete this you have to delete sub items first!']);
+        }
         $subCategory->delete();
         return response(['status' => 'success', 'message' => 'Deleted Successfully']);
     }
